1.碎片图像复原----第一版/Debug/requirements.txt
pillow
numpy
pyqt5
sklearn